Refactor variable type in CPU utility percentage methods
In the CPU class methods currentUtility_Percentage and currentThreadUtility_Percentage, the type of the query variable was changed from std::string& to const std::string&. This was done to prevent inadvertent modification of the query string in those methods, improving code safety. The 'const int&' parameter in currentThreadUtility_Percentage method was also refactored to 'int' for simplicity since integers are cheap to copy.