Refactor CMakeLists for code standard requirement
The project CMakeLists.txt file was refactored so that the 'CMAKE_CXX_STANDARD_REQUIRED' directive is applied regardless of 'NO_OCL' check. Removed redundant flag settings for RELEASE and DEBUG mode. This simplifies the structure of the file and makes it clearer that CXX standard requirement applies universally, and also avoids spurious compiler warnings about "unknown flag -O3" on MSVC and clang-cl.