Update C++ Standard and Make Conditions Constexpr
The C++ standard used in the NO_OCL branch of CMakeLists.txt was updated from C++11 to C++17 to use newer language features. Also, in the WMIwrapper.h, conditions in multiple `if` statements were changed to `if constexpr` to resolve compiler warnings. This change ensures the evaluated condition is a constant expression while limiting type checking to the applicable branch only, resulting in cleaner, more efficient code.