Replace unsafe methods with secure alternatives
The changes in this commit replace getenv and wcstombs with safer alternatives, namely _dupenv_s and wcstombs_s respectively. This improves security by ensuring buffer sizes are accurately accounted for, reducing the risk of overflows and crashes. In particular, _dupenv_s is used to safely get the value of the HOME environment variable in PCIMapper.cpp, and wcstombs_s is used for safely converting wide strings to regular strings in stringutils.h.