Lower C++ standard requirement and remove constexpr
Change 'CMAKE_CXX_STANDARD' in CMakeLists.txt from C++17 to C++11 when 'NO_OCL' is set. Also, remove use of 'constexpr' in WMIwrapper.h

The C++ standard requirement was lowered to ensure compatibility with some older compilers. The use of 'constexpr' in if conditionals was removed for compatibility with the lower standard requirement. Now the build should successfully compile on older systems.