Remove redundant code
The excessive getter functions in *hwinfo* component have been removed and replaced by direct member access inside the classes. The redundant files in the *source* directory have been deleted. The motivation for this change comes from the efforts to revise and simplify the codebase, increasing the readability and maintainability of the code. The deletion of unnecessary files contributed to freeing up disk space and decluttering the project structure. This commit's action should not affect the program's functionality or performance.