Refactor apple hwinfo structure
The structure of the hardware info for Apple systems has been updated. The import style for 'platform.h' and 'mainboard.h' was changed for simplicity. The functions 'getVendor', 'getName', 'getVersion', and 'getSerialNumber' have been moved inside the 'mainboard' namespace for a more coherent organization. The 'getRegularClockSpeed_kHz' function inside 'cpu.h' has been temporarily commented out as it is not currently being used.